Challenge 1 all methods completed ProgramUI

+++ Challenge_1.ProgramUI/ProgramUI.cs
@@ -21,6 +21,7 @@ namespace Challenge_1.ProgramUI
             bool keeprunning = true;
             while (keeprunning)
             {
+                Console.Clear();
                 Console.WriteLine("Welcome\n" +
                     "What would you like to do?");
 
@@ -28,7 +29,7 @@ namespace Challenge_1.ProgramUI
                     "1. Create a new menu item\n" +
                     "2. View all menu items\n" +
                     "3. View menu item by meal number\n" +
-                    "4. Update an esisting menu item\n" +
+                    "4. Update an existing menu item\n" +
                     "5. Delete an existing menu item\n" +
                     "6. Exit");
 
@@ -90,22 +91,27 @@ namespace Challenge_1.ProgramUI
             {
                 Console.WriteLine("The new menu item could not be added. Please try again.");
             }
+            Console.ReadKey();
         }
 
         private void DisplayAllMenuItems()
         {
             Console.Clear();
             List<MenuItem> allMenuItems = _repo.GetMenuItems();
-            if (allMenuItems != null)
+            if (allMenuItems.Count != 0)
             {
                 foreach (MenuItem menuItem in allMenuItems)
                 {
                     Console.WriteLine($"Menu Number: {menuItem.MealNumber}\n" +
                         $"{menuItem.MealName}\n" +
-                        $"Price: {menuItem.MealPrice}");
+                        $"Price: ${menuItem.MealPrice}");
                 }
             }
-            else Console.WriteLine("There are no menu items to display");
+            else
+            {
+                Console.WriteLine("There are no menu items to display");
+            }
+            Console.ReadKey();
         }
 
         private void DisplayMenuItemByNumber()
@@ -120,56 +126,84 @@ namespace Challenge_1.ProgramUI
                     $"{menuItemToDisplay.MealName}\n" +
                     $"Description: {menuItemToDisplay.Description}\n" +
                     $"Ingredients: {menuItemToDisplay.IngredientList}\n" +
-                    $"Price: {menuItemToDisplay.MealPrice}");
+                    $"Price: ${menuItemToDisplay.MealPrice}");
             }
-            else Console.WriteLine("There is no menu item by that number");
+            else
+            {
+                Console.WriteLine("There is no menu item by that number");
+            }
+            Console.ReadKey();
         }
 
         private void UpdateExistingMenuItem()
         {
             Console.Clear();
-            DisplayAllMenuItems();
-            Console.WriteLine("Please enter the menu number of the menu item that you would like to update.");
-
-            int menuNumber = Convert.ToInt32(Console.ReadLine());
-
-            MenuItem updateMenuItem = new MenuItem();
-
-            Console.WriteLine("Please enter a new menu number");
-            updateMenuItem.MealNumber = Convert.ToInt32(Console.ReadLine());
-
-            Console.WriteLine("Please enter the name of the meal");
-            updateMenuItem.MealName = Console.ReadLine();
-
-            Console.WriteLine("Please enter a meal description");
-            updateMenuItem.Description = Console.ReadLine();
-
-            Console.WriteLine("Please enter the ingredient of the meal");
-            updateMenuItem.IngredientList = Console.ReadLine();
-
-            Console.WriteLine("Finally, enter the price of the updated meal");
-            updateMenuItem.MealPrice = Convert.ToDouble(Console.ReadLine());
-
-            bool wasUpdated = _repo.UpdateExistingMenuItem(menuNumber, updateMenuItem);
-            if (wasUpdated)
+            List<MenuItem> allMenuItems = _repo.GetMenuItems();
+            if (allMenuItems.Count > 0)
             {
-                Console.WriteLine("The menu item was successfully updated");
+                DisplayAllMenuItems();
+                Console.WriteLine("Please enter the menu number of the menu item that you would like to update.");
+
+                int menuNumber = Convert.ToInt32(Console.ReadLine());
+
+                MenuItem updateMenuItem = new MenuItem();
+
+                Console.WriteLine("Please enter a new menu number");
+                updateMenuItem.MealNumber = Convert.ToInt32(Console.ReadLine());
+
+                Console.WriteLine("Please enter the name of the meal");
+                updateMenuItem.MealName = Console.ReadLine();
+
+                Console.WriteLine("Please enter a meal description");
+                updateMenuItem.Description = Console.ReadLine();
+
+                Console.WriteLine("Please enter the ingredients of the meal");
+                updateMenuItem.IngredientList = Console.ReadLine();
+
+                Console.WriteLine("Finally, enter the price of the updated meal");
+                updateMenuItem.MealPrice = Convert.ToDouble(Console.ReadLine());
+
+                bool wasUpdated = _repo.UpdateExistingMenuItem(menuNumber, updateMenuItem);
+                if (wasUpdated)
+                {
+                    Console.WriteLine("The menu item was successfully updated");
+                }
+                else
+                {
+                    Console.WriteLine("The menu item could not be updated");
+                }
             }
-            else Console.WriteLine("The menu item could not be updated");
+            else
+            {
+                Console.WriteLine("There are no menu items to update");
+            }
+            Console.ReadKey();
         }
 
         private void DeleteExistingMenuItem()
         {
             Console.Clear();
-            DisplayAllMenuItems();
-            Console.WriteLine("Please select a menu item number to delete");
-
-            bool wasDeleted = _repo.DeleteExistingMenuItem(Convert.ToInt32(Console.ReadLine()));
-            if (wasDeleted)
+            List<MenuItem> allMenuItems = _repo.GetMenuItems();
+            if (allMenuItems.Count > 0)
             {
-                Console.WriteLine("The menu item was successfully deleted");
+                DisplayAllMenuItems();
+                Console.WriteLine("Please select a menu item number to delete");
+
+                bool wasDeleted = _repo.DeleteExistingMenuItem(Convert.ToInt32(Console.ReadLine()));
+                if (wasDeleted)
+                {
+                    Console.WriteLine("The menu item was successfully deleted");
+                }
+                else
+                {
+                    Console.WriteLine("The menu item could not be deleted");
+                }
             }
-            else Console.WriteLine("The menu item could not be deleted");
+            else
+            {
+                Console.WriteLine("There is nothing to delete");
+            }
+            Console.ReadKey();
         }
     }
 }
